Additional Files
newDB.mv.db is the database file I am using currently.

pom.xml are the dependencies for the code.

sql_stuff is just for the creating the database if you don't want to use
the db file. These files will help you make a db really quick. Not
necessary for running the code.